tipe data

+++ MasterOnline/Controllers/StokControllerJob.cs
@@ -2106,25 +2106,25 @@ namespace MasterOnline.Controllers
 
         public class TokpedGetProductInfoBasic
         {
-            public int productID { get; set; }
-            public int shopID { get; set; }
+            public long productID { get; set; }
+            public long shopID { get; set; }
             public int status { get; set; }
             public string name { get; set; }
             public int condition { get; set; }
-            public int childCategoryID { get; set; }
+            public long childCategoryID { get; set; }
         }
 
         public class TokpedGetProductInfoPrice
         {
-            public int value { get; set; }
+            public long value { get; set; }
             public int currency { get; set; }
-            public int LastUpdateUnix { get; set; }
-            public int idr { get; set; }
+            public long LastUpdateUnix { get; set; }
+            public long idr { get; set; }
         }
 
         public class TokpedGetProductInfoWeight
         {
-            public int value { get; set; }
+            public double value { get; set; }
             public int unit { get; set; }
         }
 
@@ -2137,9 +2137,9 @@ namespace MasterOnline.Controllers
 
         public class TokpedGetProductInfoVariant
         {
-            public int parentID { get; set; }
+            public long parentID { get; set; }
             public bool isVariant { get; set; }
-            public int[] childrenID { get; set; }
+            public long[] childrenID { get; set; }
         }
 
         public class TokpedGetProductInfoMenu
@@ -2155,7 +2155,7 @@ namespace MasterOnline.Controllers
         public class TokpedGetProductInfoExtraattribute
         {
             public int minOrder { get; set; }
-            public int lastUpdateCategory { get; set; }
+            public long lastUpdateCategory { get; set; }
             public bool isEligibleCOD { get; set; }
         }
 
@@ -2186,7 +2186,7 @@ namespace MasterOnline.Controllers
 
         public class TokpedGetProductInfoCategorytree
         {
-            public int id { get; set; }
+            public long id { get; set; }
             public string name { get; set; }
             public string title { get; set; }
             public string breadcrumbURL { get; set; }
@@ -2194,7 +2194,7 @@ namespace MasterOnline.Controllers
 
         public class TokpedGetProductInfoPicture
         {
-            public int picID { get; set; }
+            public long picID { get; set; }
             public string fileName { get; set; }
             public string filePath { get; set; }
             public int status { get; set; }
@@ -2207,18 +2207,18 @@ namespace MasterOnline.Controllers
 
         public class TokpedGetProductInfoWarehouse
         {
-            public int productID { get; set; }
-            public int warehouseID { get; set; }
+            public long productID { get; set; }
+            public long warehouseID { get; set; }
             public TokpedGetProductInfoPrice1 price { get; set; }
             public TokpedGetProductInfoStock1 stock { get; set; }
         }
 
         public class TokpedGetProductInfoPrice1
         {
-            public int value { get; set; }
+            public long value { get; set; }
             public int currency { get; set; }
-            public int LastUpdateUnix { get; set; }
-            public int idr { get; set; }
+            public long LastUpdateUnix { get; set; }
+            public long idr { get; set; }
         }
 
         public class TokpedGetProductInfoStock1
